requirements.txt
tensorflow
tqdm
scikit-learn
opencv-python
streamlit
faiss-cpu